Rename three files:
app/src/ee/ioc/phon/android/speak/VoiceImeService.java -> app/src/ee/ioc/phon/android/speak/service/SpeechInputMethodService.java
app/src/ee/ioc/phon/android/speak/VoiceImeView.java -> app/src/ee/ioc/phon/android/speak/view/SpeechInputView.java
app/src/ee/ioc/phon/android/speak/MicButton.java -> app/src/ee/ioc/phon/android/speak/view/MicButton.java
