+++ app/AndroidManifest.xml
@@ -270,7 +270,7 @@ Also, failed to set the launchMode from the resource, INSTALL_PARSE_FAILED_UNEXP
         </service>
 
         <service
-            android:name="VoiceImeService"
+            android:name=".service.SpeechInputMethodService"
             android:label="@string/labelIme"
             android:permission="android.permission.BIND_INPUT_METHOD">
             <intent-filter>
