+++ app/res/layout-land/voice_ime_view.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<ee.ioc.phon.android.speak.VoiceImeView xmlns:android="http://schemas.android.com/apk/res/android"
+<ee.ioc.phon.android.speak.view.SpeechInputView xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_alignParentBottom="true"
@@ -82,7 +82,7 @@
             android:textColor="@color/grey200"
             android:textSize="@dimen/textSize10sp" />
 
-        <ee.ioc.phon.android.speak.MicButton
+        <ee.ioc.phon.android.speak.view.MicButton
             android:id="@+id/bImeStartStop"
             android:layout_width="@dimen/layoutHeightButtonMicrophone"
             android:layout_height="@dimen/layoutHeightButtonMicrophone"
@@ -95,4 +95,4 @@
 
     </RelativeLayout>
 
-</ee.ioc.phon.android.speak.VoiceImeView>
+</ee.ioc.phon.android.speak.view.SpeechInputView>
